Рефакторинг
- выносит повторяющиеся данные в модуль с константами
- выносит вспомогательные функции в модуль с утилитами
- выносит общие для компонентов функции в модуль с утилитами
