+++ src/mock/task.js
@@ -1,9 +1,5 @@
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+import {COLORS} from "../const.js";
+import {getRandomInteger} from "../utils.js";
 
 const generateDescription = () => {
   const descriptions = [
@@ -47,10 +43,9 @@ const generateRepeating = () => {
 };
 
 const getRandomColor = () => {
-  const colors = [`black`, `yellow`, `blue`, `green`, `pink`];
-  const randomIndex = getRandomInteger(0, colors.length - 1);
+  const randomIndex = getRandomInteger(0, COLORS.length - 1);
 
-  return colors[randomIndex];
+  return COLORS[randomIndex];
 };
 
 export const generateTask = () => {
